pyproject.toml
[tool.poetry]
name = "flask34234"
version = "0.1.0"
description = ""
authors = ["X0re4ik <xoore4ik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.2"
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
